moving functions to helpers.
