+++ client/containers/workflow-list/helpers/is-route-range-valid.js
@@ -19,13 +19,18 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-export { default as getCriteria } from './get-criteria';
-export { default as getFetchWorkflowListUrl } from './get-fetch-workflow-list-url';
-export { default as getFilterBy } from './get-filter-by';
-export { default as getFilterModeButtonLabel } from './get-filter-mode-button-label';
-export { default as getFormattedResults } from './get-formatted-results';
-export { default as getMinStartDate } from './get-min-start-date';
-export { default as getState } from './get-state';
-export { default as getStatus } from './get-status';
-export { default as isRangeValid } from './is-range-valid';
-export { default as isRouteRangeValid } from './is-route-range-valid';
+import isRangeValid from './is-range-valid';
+
+const isRouteRangeValid = ({ endTime, minStartDate, range, startTime }) => {
+  if (range) {
+    return isRangeValid(range, minStartDate);
+  }
+
+  if (startTime && endTime) {
+    return isRangeValid({ endTime, startTime }, minStartDate);
+  }
+
+  return false;
+};
+
+export default isRouteRangeValid;
